==> src/3-modeling-prep.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
import datetime
from sklearn.model_selection import train_test_split
from feature_engineering import feature_engineering
# from category_encoders import TargetEncoder

# How to use this notebook

Configure global variables `SEASON`, `TRAIN_TEAM_SEASON`, and `TEST_TEAM_SEASON` to get the desired datasets configuration.

The train-set and test-test produced here is used in `4-linear-models.ipynb` and `5-random-forrest.ipynb` notebooks

df = pd.read_csv('../data/data2.csv')

SEASON = '2009-10'

# TRAIN_TEAM_SEASONS = ['2000-01', '2001-02', '2002-03']
# TRAIN_TEAM_SEASONS = ['2001-02', '2002-03', '2003-04']
# TRAIN_TEAM_SEASONS = ['2002-03', '2003-04', '2004-05']
TRAIN_TEAM_SEASONS = ['1999-00', '2000-01', '2001-02', '2002-03', '2003-04']
TEST_TEAM_SEASONS = ['2005-06']
TEAM = 'PHX'

df.head()

***
## Removing Redundant Columns
There are some columns that have very high corelations with some other columns and some columns that

Removing columns:
- team_name, team_id - These have only a single value, so it doesn't add any value to the prediction
- matchup - It has 1 to 1 corelation with the column "opponent"

def remove_columns(df, remove_season=False):
    columns_to_drop = ['game_event_id', 'lat', 'lon', 'team_name', 'team_id', 'game_id', 'shot_id', 'game_date',
                       'season', 'loc_x', 'loc_y']
    df = df.drop(labels=columns_to_drop, axis='columns')
    if(type == True):
        train.drop(['month', 'playoffs'], axis=1, inplace=True)
    print(f"Total columns removed: {len(columns_to_drop)}")
    return df

#### Helper function

def remove_na_rows(df: pd.DataFrame):
    """
    :param pd.DataFrame df:
    :return pd.DataFrame
    """
    num_rows_before = len(df)
    print(f'Number of rows: {num_rows_before}')
    df = df.dropna()
    num_rows_after = len(df)
    print(f'Number of rows after dropping missing values: {num_rows_after}')
    removed_rows = num_rows_before - num_rows_after
    print(f'Percentage of data removed: {100 * (removed_rows / num_rows_before)} %')
    return df

# Preparing for modeling
From the analyzed and preprocessed data, we continue the pre-processing in order to prepare it for the modeling phase

def target_encoding(df, col_to_encode, target_col, is_trainset=True, encoder=None):
    """
    For trainset:
        Performs target-encoding on the specified column. That is to convert categorical column
        to a numerical column. Drops the col_to_encode from the returned DataFrame.
    For test-set:
        Performs target-encoding on the specified column using the encoding given from the trainset.
        Drops the col_to_encode from the returned DataFrame.

    Args:
        df (DataFrame):
        col_to_encode (str): Column to be converted to numerical one.
        target_col (str): Target column of the data.
        is_trainset (bool): Tells if the df is training-set or testing-set
        encoding (dict): encoder that tells how to map the categorical values.

    Returns: (pd.DataFrame, dict)
        DataFrame: DataFrame containing the new <col_to_encode>_te column.
        dict: is the encoder.
    """
    if is_trainset:
        encoder = df.groupby(col_to_encode)[target_col].mean().reset_index()
        encoder = encoder.rename(columns={target_col: f"{col_to_encode}_te"})
        df = df.merge(right=encoder, how='left', on=col_to_encode)
        df.drop(col_to_encode, axis=1, inplace=True)
        
    elif is_trainset == False:
        assert encoder is not None
        
        df = df.merge(right=encoder, how='left', on=col_to_encode)
        df.drop(col_to_encode, axis=1, inplace=True)
        
    return df, encoder

def encode_and_bind(original_dataframe, feature_to_encode):
    dummies = pd.get_dummies(original_dataframe[[feature_to_encode]])
    res = pd.concat([original_dataframe, dummies], axis=1)
    res = res.drop([feature_to_encode], axis=1)
    return res

def one_hot_encoding(df):
#     df_new = df.copy(deep=True)
#     features_to_encode = ['combined_shot_type', 'season', 'shot_type', 'shot_zone_area', 'shot_zone_basic', 'shot_zone_range', 'opponent', 'action_type', 'matchup']
    features_to_encode = ['combined_shot_type', 'shot_type', 'shot_zone_area', 'matchup']
    for feature in features_to_encode:
        df = encode_and_bind(df, feature)

    return df

def ordinal_encoder(df: pd.DataFrame, col_to_encode, mapper: dict) -> pd.DataFrame:
    """
    Performs ordinal encoding using the given mapper. It replaces the col_to_encode's values in
    DataFrame.
    Args:
        df (DataFrame):
        col_to_encode (str): column that needs to be encoded
        mapper (dict): contains the mapping for the column in df. 
            e.g. "col_val" -> 1

    Returns:
        DataFrame: returns the DataFrame with mapped values in col: col_to_encode
    """
    df[col_to_encode] = df[col_to_encode].apply(lambda x: mapper[x])
    return df

### Encoding

shot_zone_basic_mapper = {    
    "Restricted Area": 0,
    "In The Paint (Non-RA)": 1,
    "Mid-Range": 2,
    "Right Corner 3": 3,
    "Above the Break 3": 4,
    "Left Corner 3": 5,
    "Backcourt": 6
}
shot_zone_range_mapper = {
    "Less Than 8 ft.": 0,
    "8-16 ft.": 1,
    "16-24 ft.": 2,
    "24+ ft.": 3,
    "Back Court Shot": 4
}

#### Splitting features & label (target)

def split_x_y(df, target):
    y = df.loc[:, [target]]
    x = df.drop([target], axis=1)
    return x, y

# Regular (train), playoff (test)

def split_regular_playoff(df, season, train=None, test=None, validation=None, n_way_split=False):

    # if(train + test + validation != 1):
    #     print("Error dataset split")
    #     return None

    df_validaiton = None

    df_season = df[df['season'] == season]
    isPlayoffs = df_season.playoffs == 1

    df_train = df_season[~isPlayoffs]
    df_test  = df_season[isPlayoffs]

    print(f"Size of training dataset {len(df_train)}" )
    print(f"Size of test dataset {len(df_test)}" )
    print(f"Size of train vs test ratio {len(df_train) / (len(df_train) + len(df_test)) }" )

    if (n_way_split == True):
        train = split_x_y(df_train, 'shot_made_flag')
        test = split_x_y(df_test, 'shot_made_flag')
        return train[0], train[1], test[0], test[1]
    else:
        return df_train, df_test

train, test = split_regular_playoff(df, SEASON)

train = feature_engineering(train)
train = remove_columns(train, remove_season=True)

test = feature_engineering(test)
test = remove_columns(test, remove_season=True)

train, action_type_encoder = target_encoding(train, 'action_type', 'shot_made_flag')
train, opponent_type_encoder = target_encoding(train, 'opponent', 'shot_made_flag')
train = one_hot_encoding(train)
train = ordinal_encoder(train, "shot_zone_basic", shot_zone_basic_mapper)
train = ordinal_encoder(train, "shot_zone_range", shot_zone_range_mapper)

test, _ = target_encoding(test, 'action_type', 'shot_made_flag', is_trainset=False, encoder=action_type_encoder)
test, _ = target_encoding(test, 'opponent', 'shot_made_flag', is_trainset=False, encoder=opponent_type_encoder)
test = one_hot_encoding(test)
test = ordinal_encoder(test, "shot_zone_basic", shot_zone_basic_mapper)
test = ordinal_encoder(test, "shot_zone_range", shot_zone_range_mapper)

test = remove_na_rows(test)

## missing features in train2 and test2

print(set(train.columns.values) == (set(test.columns.values)))

print(set(test.columns.values).difference(set(train.columns.values)))
print(set(train.columns.values).difference(set(test.columns.values)))

for train_features in set(test.columns.values).difference(set(train.columns.values)):
    train[train_features] = 0

for test_features in set(train.columns.values).difference(set(test.columns.values)):
    test[test_features] = 0

print(set(train.columns.values) == (set(test.columns.values)))

print(f"Train Shape: {train.shape}")
print(f"Test Shape: {test.shape}")

train.to_csv('../data/train_season.csv', index=False)
test.to_csv('../data/test_season.csv', index=False)

# Pick one team, train seasons before, test seasons after

In this version of dataset we get all the matches played against a single team in one season e.g. season `2008-09` 
and then test the model against all the matches played against same team in the next season e.g. season `2009-10`

def split_team_season(df, opponent,  season_train: list, season_test: list,
                      train=None, test=None, validation=None):

    # if(train + test + validation != 1):
    #     print("Error dataset split")
    #     return None

    df_train = None
    df_test = None
    df_validaiton = None

    if season_train == [] or season_test == []:
        print("Empty seasons")
        return None

    df_train = df[(df['season'].isin(season_train))  & (df['opponent'] == opponent) ]
    df_test = df[(df['season'].isin(season_test)) & (df['opponent'] == opponent) ]
    print(f"Testing vs {opponent}\ntraining: {season_train}\ntesting: {season_test}" )
    print(f"Size of training dataset {len(df_train)}" )
    print(f"Size of test dataset {len(df_test)}" )
    print(f"Size of train vs test ratio {len(df_train) / (len(df_train) + len(df_test)) }" )

    return df_train, df_test

train2, test2 = split_team_season(df, opponent=TEAM, season_train=TRAIN_TEAM_SEASONS, season_test=TEST_TEAM_SEASONS)

train2.to_csv("temp.csv", index=False)

train2 = feature_engineering(train2)
train2 = remove_columns(train2, remove_season=False)

test2 = feature_engineering(test2)
test2 = remove_columns(test2, remove_season=False)

train2, action_type_encoder = target_encoding(train2, 'action_type', 'shot_made_flag')
train2, opponent_type_encoder = target_encoding(train2, 'opponent', 'shot_made_flag')
train2 = one_hot_encoding(train2)
train2 = ordinal_encoder(train2, "shot_zone_basic", shot_zone_basic_mapper)
train2 = ordinal_encoder(train2, "shot_zone_range", shot_zone_range_mapper)

test2, _ = target_encoding(test2, 'action_type', 'shot_made_flag', is_trainset=False, encoder=action_type_encoder)
test2, _ = target_encoding(test2, 'opponent', 'shot_made_flag', is_trainset=False, encoder=opponent_type_encoder)
test2 = one_hot_encoding(test2)
test2 = ordinal_encoder(test2, "shot_zone_basic", shot_zone_basic_mapper)
test2 = ordinal_encoder(test2, "shot_zone_range", shot_zone_range_mapper)

test2 = remove_na_rows(test2)

print(f"Rows in training-set: {len(train2)}")
print(f"Rows in test-set: {len(test2)}")

## missing features in train2 and test2

It could happen that the train and test data have different number of columns when converting some categorical feature to numeric, therefore in this part we are balancing both datasets so that they have same number of columns. 

print(set(train2.columns.values) == (set(test2.columns.values)))

print(set(test2.columns.values).difference(set(train2.columns.values)))
print(set(train2.columns.values).difference(set(test2.columns.values)))

for train_features in set(test2.columns.values).difference(set(train2.columns.values)):
    train2[train_features] = 0

for test_features in set(train2.columns.values).difference(set(test2.columns.values)):
    test2[test_features] = 0

print(set(train2.columns.values) == (set(test2.columns.values)))

train2.columns.values == 'combined_shot_type_Tip Shot'

# train2.head()
df.action_type.unique()

train2.to_csv('../data/train_team.csv', index=False)
test2.to_csv('../data/test_team_05-06.csv', index=False)
