==> src/4-linear-models.ipynb <==
import pandas as pd
import numpy as np
import pprint
from sklearn.linear_model import LinearRegression, Ridge, RidgeCV, Lasso, LassoCV, RidgeClassifierCV
from sklearn.metrics import mean_squared_error, r2_score, accuracy_score, f1_score, classification_report

from sklearn.model_selection import train_test_split,  KFold, cross_val_score
from sklearn.neural_network import MLPRegressor, MLPClassifier

# from IPython.core.interactiveshell import InteractiveShell
# InteractiveShell.ast_node_interactivity = "all"

pd.set_option("display.max_columns", None)

# How to use this Notebook

Change `DATASET` and `VAL_ON_END` variable to test the dataset you want.
For explanation of `SEASON` and `TEAM` read notebook `3-modeling-prep.ipynb`

DATASET = 'SEASON'
# DATASET = 'TEAM'
VAL_ON_END = True  # False = Start. False would mean validate on starting season's data

if DATASET == 'TEAM':
    train = pd.read_csv('../data/train_team.csv')
    test = pd.read_csv('../data/test_team.csv')
#     test  = pd.read_csv('../data/test_team_15-16.csv')

if DATASET == 'SEASON':
    train = pd.read_csv('../data/train_season.csv')
    test = pd.read_csv('../data/test_season.csv')

if VAL_ON_END:
    # To validate on data of the last season's, in the training-set
    train, validation = train_test_split(train, train_size=0.8, shuffle=False)
else:
    # To validate on data of the starting season's, in the training-set
    validation, train = train_test_split(train, train_size=0.2, shuffle=False)

target_column = "shot_made_flag"  # y_column_name

def confusion(true, pred):
    """
    Function for pretty printing confusion matrices
    """
    pred = pd.Series(pred)
    true = pd.Series(true)
    
    true.name = 'target'
    pred.name = 'predicted'
    cm = pd.crosstab(true.reset_index(drop=True), pred.reset_index(drop=True))
    print(cm)
    cm = cm[cm.index]
    return cm

def compute_metrics(y_true,y_pred):
    accuracy = accuracy_score(y_true,y_pred)
    f1_score_1 = f1_score(y_true,y_pred,average='binary',pos_label=1)
    f1_score_0 = f1_score(y_true,y_pred,average='binary',pos_label=0)
    f1_score_macro = f1_score(y_true,y_pred,average='macro')
    return [accuracy,f1_score_1,f1_score_0,f1_score_macro]

train_res = pd.DataFrame(columns=['Accuracy', 'F1-score (class 1)', 'F1-score (class 0)', 'F1-score (macro avg)']) 
results = pd.DataFrame(columns=['Accuracy', 'F1-score (class 1)', 'F1-score (class 0)', 'F1-score (macro avg)'])

def split_x_y(df, target):
    y = df.loc[:, [target]]
    x = df.drop([target], axis=1)
    return x, y

def split_regular_playoff(df, season, train=None, test=None, validation=None):

    # if(train + test + validation != 1):
    #     print("Error dataset split")
    #     return None

    df_validaiton = None

    df_season = df[df['season'] == season]
    isPlayoffs = df_season.playoffs == 1

    df_train = df_season[~isPlayoffs]
    df_test  = df_season[isPlayoffs]

    print(f"Size of training dataset {len(df_train)}" )
    print(f"Size of test dataset {len(df_test)}" )
    print(f"Size of train vs test ratio {len(df_train) / (len(df_train) + len(df_test)) }" )

    return df_train, df_test
#     train = split_x_y(df_train, 'shot_made_flag')
#     test = split_x_y(df_test, 'shot_made_flag')
#     return train[0], train[1], test[0], test[1]
#     return train, test

# x_train, y_train, x_test, y_test = split_regular_playoff(df, '2010-11')
# train, test = split_regular_playoff(df, '2010-11')

## Split X & Y

x_train, y_train = split_x_y(train, target_column)
x_test, y_test = split_x_y(test, target_column)
x_val, y_val = split_x_y(validation, target_column)

print(f"Shape in training-set: {x_train.shape}")
print(f"Validation-set Shape: {x_val.shape}")
print(f"Test-set Shape: {x_test.shape}")

lambdas = [1e-10, 1e-9, 1e-8, 1e-6, 1e-5, 1e-4, 1e-3, 1e-2, 0.1, 0.5, 1, 5, 10, 50,100]

# Linear Regression

x_train.describe()

reg = LinearRegression()
a = reg.fit(x_train, y_train)
reg.score(x_train, y_train)

#### train data

y_train_pred = reg.predict(x_train)
y_train_pred = y_train_pred >= 0.5
y_train_pred = np.multiply(y_train_pred, 1)
print(f"Accuracy on training-set: {accuracy_score(y_train.values, y_train_pred)}")
train_res.loc['Linear',:] = compute_metrics(y_train.values, y_train_pred)
y_val_pred = reg.predict(x_val)
y_val_pred = y_val_pred >= 0.5
y_val_pred = np.multiply(y_val_pred, 1)
print(f"Accuracy on Validation-set: {accuracy_score(y_val.values, y_val_pred)}")
results.loc['Linear',:] = compute_metrics(y_val.values, y_val_pred)
results

# Lasso Regression

lasso_cv = LassoCV(alphas=lambdas, cv=5, max_iter=100_000)
lasso_cv.fit(x_train.values, y_train.values.flatten())
lasso_r2 =  np.mean(cross_val_score(lasso_cv, x_train.values, y_train.values.flatten()))
print('Best lambda:', lasso_cv.alpha_, 'R2 score:',lasso_r2)

y_train_pred = lasso_cv.predict(x_train.values)
y_train_pred = y_train_pred >= 0.5
y_train_pred = np.multiply(y_train_pred, 1)
print(f"Accuracy on training-set: {accuracy_score(y_train.values, y_train_pred):.5f}")
train_res.loc['Lasso',:] = compute_metrics(y_train.values, y_train_pred)
y_val_pred = lasso_cv.predict(x_val.values)
y_val_pred = y_val_pred >= 0.5
y_val_pred = np.multiply(y_val_pred, 1)
print(f"Accuracy on Validation-set: {accuracy_score(y_val.values, y_val_pred)}")
results.loc['Lasso',:] = compute_metrics(y_val.values, y_val_pred)
results

# Ridge Regression

best_cv = None
max_acc = -10
# Loop to find the best CV value, I've seen a good alpha is dependant on CV
for cv in [2, 3, 4, 5, 6, 7, 8]:
    ridge = RidgeClassifierCV(alphas=lambdas, cv=cv)
    ridge.fit(x_train.values, y_train.values.flatten())
#     print('Best lambda:', ridge.alpha_, 'score:', ridge.score(x_train.values, y_train.values.flatten()))
    y_train_pred = ridge.predict(x_train.values)
#     print(f"Accuracy on training-set: {accuracy_score(y_train.values, y_train_pred):.5f}")
    ry_val_pred = ridge.predict(x_val.values)
    ry_val_pred = ry_val_pred >= 0.5
    ry_val_pred = np.multiply(ry_val_pred, 1)
#     print(f"Accuracy on validation-set: {accuracy_score(y_val.values, ry_val_pred)}")
    if max_acc < accuracy_score(y_val.values, ry_val_pred):
        max_acc = accuracy_score(y_val.values, ry_val_pred)
        best_cv = cv

ridge = RidgeClassifierCV(alphas=lambdas, cv=best_cv)
ridge.fit(x_train.values, y_train.values.flatten())
print('Best lambda:', ridge.alpha_, 'score:', ridge.score(x_train.values, y_train.values.flatten()))

y_train_pred = ridge.predict(x_train.values)
print(f"Accuracy on training-set: {accuracy_score(y_train.values, y_train_pred):.5f}")
ry_val_pred = ridge.predict(x_val.values)
print(f"Accuracy on validation-set: {accuracy_score(y_val.values, ry_val_pred)}")

train_res.loc['Ridge',:] = compute_metrics(y_train.values, y_train_pred)
results.loc['Ridge',:] = compute_metrics(y_val.values, ry_val_pred)

train_res

results

# Test Data
x_test.shape

x_val.shape

x_test, y_test
print(f"Rows in test-set: {len(x_test)}")
ry_test_pred = ridge.predict(x_test.values)
print(f"Accuracy on Test-set: {accuracy_score(y_test.values, ry_test_pred)}")
